Add support for loading dataset from local ZIP file
This commit adds support for loading the dataset from a local ZIP file. The `load_dataset` function now accepts a `zip_path` parameter, which specifies the path to the local ZIP file. If the `zip_path` is provided, the ZIP file will be loaded from that path instead of downloading it from the Figshare API. The `Dataset` class has been updated to handle the new `zip_path` parameter and load the ZIP file from the specified path.